Move the worker into his own file and import analyze function
